refactor: extract useClickOutside and useTabWorkspaceActions hooks
- R1: Extract useClickOutside hook from StatusBar and TabContextMenu
  to deduplicate mousedown click-outside detection pattern
- R2: Extract useTabWorkspaceActions hook from App.tsx containing
  handleSelectWorkspace, handleSelectTab, handleCloseTab, handleAddTab,
  handleReorderTabs (with O4 fix: workspace-only reorder), handleContextAction,
  handleMiddleClick, and contextMenu state management
- App.tsx reduced to store subscriptions, derived state, and JSX
