+++ spa/src/App.tsx
@@ -1,5 +1,5 @@
 // spa/src/App.tsx — v2 重構：wouter Router + Tab/Pane model
-import { useEffect, useState, useCallback } from 'react'
+import { useEffect, useState } from 'react'
 import { Router } from 'wouter'
 import { ActivityBar } from './components/ActivityBar'
 import { TabBar } from './components/TabBar'
@@ -15,14 +15,13 @@ import { useHistoryStore } from './stores/useHistoryStore'
 import { useRelayWsManager } from './hooks/useRelayWsManager'
 import { useSessionEventWs } from './hooks/useSessionEventWs'
 import { useRouteSync } from './hooks/useRouteSync'
-import { createTab, isStandaloneTab } from './types/tab'
-import { getPrimaryPane } from './lib/pane-tree'
-import { TabContextMenu, type ContextMenuAction } from './components/TabContextMenu'
+import { useTabWorkspaceActions } from './hooks/useTabWorkspaceActions'
+import { isStandaloneTab } from './types/tab'
+import { TabContextMenu } from './components/TabContextMenu'
 import type { Tab } from './types/tab'
 
 export default function App() {
   const [settingsOpen, setSettingsOpen] = useState(false)
-  const [contextMenu, setContextMenu] = useState<{ tab: Tab; position: { x: number; y: number } } | null>(null)
 
   // Host store (replaces hardcoded daemonBase)
   const getDaemonBase = useHostStore((s) => s.getDaemonBase)
@@ -38,19 +37,10 @@ export default function App() {
   const tabs = useTabStore((s) => s.tabs)
   const tabOrder = useTabStore((s) => s.tabOrder)
   const activeTabId = useTabStore((s) => s.activeTabId)
-  const setActiveTab = useTabStore((s) => s.setActiveTab)
-  const addTab = useTabStore((s) => s.addTab)
-  const closeTab = useTabStore((s) => s.closeTab)
-  const reorderTabs = useTabStore((s) => s.reorderTabs)
 
   // Workspace store
   const workspaces = useWorkspaceStore((s) => s.workspaces)
   const activeWorkspaceId = useWorkspaceStore((s) => s.activeWorkspaceId)
-  const setActiveWorkspace = useWorkspaceStore((s) => s.setActiveWorkspace)
-  const removeTabFromWorkspace = useWorkspaceStore((s) => s.removeTabFromWorkspace)
-  const findWorkspaceByTab = useWorkspaceStore((s) => s.findWorkspaceByTab)
-  const setWorkspaceActiveTab = useWorkspaceStore((s) => s.setWorkspaceActiveTab)
-  const reorderWorkspaceTabs = useWorkspaceStore((s) => s.reorderWorkspaceTabs)
 
   // --- Extracted hooks ---
   useRelayWsManager(wsBase)
@@ -82,48 +72,6 @@ export default function App() {
     fetchConfig(daemonBase)
   }, [fetchSessions, fetchConfig, daemonBase])
 
-  // --- Handlers ---
-
-  const handleSelectWorkspace = useCallback((wsId: string) => {
-    setActiveWorkspace(wsId)
-    const ws = workspaces.find((w) => w.id === wsId)
-    if (ws?.activeTabId) setActiveTab(ws.activeTabId)
-    else if (ws?.tabs[0]) setActiveTab(ws.tabs[0])
-  }, [workspaces, setActiveWorkspace, setActiveTab])
-
-  const handleSelectTab = useCallback((tabId: string) => {
-    setActiveTab(tabId)
-    const ws = findWorkspaceByTab(tabId)
-    if (ws) {
-      setActiveWorkspace(ws.id)
-      setWorkspaceActiveTab(ws.id, tabId)
-    }
-  }, [setActiveTab, findWorkspaceByTab, setActiveWorkspace, setWorkspaceActiveTab])
-
-  const handleCloseTab = useCallback((tabId: string) => {
-    const tab = tabs[tabId]
-    if (!tab || tab.locked) return // locked guard
-    useHistoryStore.getState().recordClose(tab, findWorkspaceByTab(tabId)?.id)
-    const ws = findWorkspaceByTab(tabId)
-    if (ws) removeTabFromWorkspace(ws.id, tabId)
-    closeTab(tabId)
-  }, [tabs, findWorkspaceByTab, removeTabFromWorkspace, closeTab])
-
-  const handleAddTab = useCallback(() => {
-    const tab = createTab({ kind: 'new-tab' })
-    addTab(tab)
-    setActiveTab(tab.id)
-    if (activeWorkspaceId) {
-      useWorkspaceStore.getState().addTabToWorkspace(activeWorkspaceId, tab.id)
-      useWorkspaceStore.getState().setWorkspaceActiveTab(activeWorkspaceId, tab.id)
-    }
-  }, [addTab, setActiveTab, activeWorkspaceId])
-
-  const handleReorderTabs = useCallback((order: string[]) => {
-    reorderTabs(order)
-    if (activeWorkspaceId) reorderWorkspaceTabs(activeWorkspaceId, order)
-  }, [reorderTabs, activeWorkspaceId, reorderWorkspaceTabs])
-
   // --- Derive visible tabs for display ---
   const activeWs = workspaces.find((w) => w.id === activeWorkspaceId)
   const visibleTabs: Tab[] = activeWs
@@ -141,52 +89,20 @@ export default function App() {
     ? [tabs[activeStandaloneTabId]].filter(Boolean)
     : visibleTabs
 
-  const handleContextMenu = useCallback((e: React.MouseEvent, tabId: string) => {
-    e.preventDefault()
-    const tab = tabs[tabId]
-    if (tab) setContextMenu({ tab, position: { x: e.clientX, y: e.clientY } })
-  }, [tabs])
-
-  const handleMiddleClick = useCallback((tabId: string) => {
-    const tab = tabs[tabId]
-    if (tab && !tab.locked) handleCloseTab(tabId)
-  }, [tabs, handleCloseTab])
-
-  const handleContextAction = useCallback((action: ContextMenuAction) => {
-    if (!contextMenu) return
-    const { tab } = contextMenu
-    const store = useTabStore.getState()
-    const primaryPaneId = getPrimaryPane(tab.layout).id
-    switch (action) {
-      case 'viewMode-terminal': store.setViewMode(tab.id, primaryPaneId, 'terminal'); break
-      case 'viewMode-stream': store.setViewMode(tab.id, primaryPaneId, 'stream'); break
-      case 'lock': case 'unlock': store.toggleLock(tab.id); break
-      case 'pin': case 'unpin': store.togglePin(tab.id); break
-      case 'close': handleCloseTab(tab.id); break
-      case 'closeOthers': {
-        const displayIds = displayTabs.map((t) => t.id)
-        const toClose = displayIds.filter((id) => id !== tab.id && !tabs[id]?.locked)
-        toClose.forEach((id) => handleCloseTab(id))
-        break
-      }
-      case 'closeRight': {
-        const displayIds = displayTabs.map((t) => t.id)
-        const idx = displayIds.indexOf(tab.id)
-        if (idx === -1) break
-        const toClose = displayIds.slice(idx + 1).filter((id) => !tabs[id]?.locked)
-        toClose.forEach((id) => handleCloseTab(id))
-        break
-      }
-    }
-  }, [contextMenu, tabs, displayTabs, handleCloseTab])
-
-  // Context menu derived state
-  const contextMenuHasRightUnlocked = (() => {
-    if (!contextMenu) return false
-    const ids = displayTabs.map((t) => t.id)
-    const idx = ids.indexOf(contextMenu.tab.id)
-    return idx !== -1 && ids.slice(idx + 1).some((id) => !tabs[id]?.locked)
-  })()
+  // --- Tab/Workspace action handlers ---
+  const {
+    contextMenu,
+    setContextMenu,
+    contextMenuHasRightUnlocked,
+    handleSelectWorkspace,
+    handleSelectTab,
+    handleCloseTab,
+    handleAddTab,
+    handleReorderTabs,
+    handleContextMenu,
+    handleMiddleClick,
+    handleContextAction,
+  } = useTabWorkspaceActions(displayTabs)
 
   return (
     <Router>
