+++ spa/src/hooks/useClickOutside.ts
@@ -1,0 +1,13 @@
+import { useEffect, type RefObject } from 'react'
+
+export function useClickOutside(ref: RefObject<HTMLElement | null>, handler: () => void) {
+  useEffect(() => {
+    function handleMouseDown(e: MouseEvent) {
+      if (ref.current && !ref.current.contains(e.target as Node)) {
+        handler()
+      }
+    }
+    document.addEventListener('mousedown', handleMouseDown)
+    return () => document.removeEventListener('mousedown', handleMouseDown)
+  }, [ref, handler])
+}
